DistortionBase in lib-builtin-effects
